[OVZC-55] 수정, 삭제 조건 추가 및 코드 수정

+++ src/main/java/kr/ac/kgu/kpserver/domain/board/Articles/ArticleService.java
@@ -1,9 +1,9 @@
 package kr.ac.kgu.kpserver.domain.board.Articles;
 
 import kr.ac.kgu.kpserver.domain.board.Articles.dto.ArticleDto;
+import kr.ac.kgu.kpserver.domain.board.Board;
 import kr.ac.kgu.kpserver.domain.board.BoardRepository;
 import kr.ac.kgu.kpserver.domain.user.User;
-import kr.ac.kgu.kpserver.domain.board.Board;
 import kr.ac.kgu.kpserver.domain.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,32 +18,42 @@ public class ArticleService {
     private final BoardRepository boardRepository;
 
     @Transactional
-    public void createdAndUpdatedArticle(Long userId, Long boardId, ArticleDto articleDto) {
+    public Long getUserId(Long userId) {
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new NotFoundException("Could not find article with id : " + userId));
+
+        return user.getId();
+    }
+
+    @Transactional
+    public void createdArticle(Long userId, Long boardId, ArticleDto articleDto) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new NotFoundException("Could not find article with id : " + userId));
         Board board = boardRepository.findById(boardId)
                 .orElseThrow(() -> new NotFoundException("Could not find article with id : " + boardId));
+        boolean canEdit = false;
         Article article =
                 new Article(articleDto.getTitle(),
-                        articleDto.getDescription(), user, board);
+                        articleDto.getDescription(), user, board, canEdit);
         articleRepository.save(article);
     }
 
     @Transactional
     public void updatedArticle(Long userId,
-                                 Long articleId,
-                                 ArticleDto articleDto) {
+                               Long articleId,
+                               ArticleDto articleDto) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new NotFoundException("Could not find article with id : " + userId));
         Article article = articleRepository.findById(articleId)
                 .orElseThrow(() -> new NotFoundException("Could not find article with id : " + articleId));
 
-        if (!user.getId().equals(article.getUser().getId())) {
-            return;
+        if (user.getId().equals(article.getUser().getId())) {
+            article.setTitle(articleDto.getTitle());
+            article.setDescription(article.getDescription());
+            boolean canEdit = true;
+            article.setCanEdit(canEdit);
+            articleRepository.save(article);
         }
-        article.setTitle(articleDto.getTitle());
-        article.setDescription(articleDto.getDescription());
-        articleRepository.save(article);
     }
 
     @Transactional(readOnly = true)
@@ -54,10 +64,16 @@ public class ArticleService {
     }
 
     @Transactional
-    public void deleteArticle(Long articleId) {
+    public void deleteArticle(Long userId, Long articleId) {
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new NotFoundException("Could not find article with id : " + userId));
         Article article = articleRepository.findById(articleId)
                 .orElseThrow(() -> new NotFoundException("Could not find article with id : " + articleId));
-        articleRepository.delete(article);
+        if (user.getId().equals(article.getUser().getId())) {
+            boolean canEdit = true;
+            article.setCanEdit(canEdit);
+            articleRepository.delete(article);
+        }
     }
 
     @Transactional
